+++ api/pom.xml
@@ -22,13 +22,13 @@
     </build>
 
     <parent>
-        <groupId>org.annotationhelper</groupId>
-        <version>1.0.0-SNAPSHOT</version>
-        <artifactId>annotationhelper-aggregator</artifactId>
+        <groupId>org.mapstruct.annotations</groupId>
+        <version>0.5.0-SNAPSHOT</version>
+        <artifactId>annotations-aggregator</artifactId>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>annotationhelper-api</artifactId>
+    <artifactId>annotations-api</artifactId>
     <packaging>jar</packaging>
     <name>Annotation Helper API</name>
