+++ api/src/main/java/org/mapstruct/annotations/Gem.java
@@ -1,4 +1,4 @@
-package org.annotationhelper;
+package org.mapstruct.annotations;
 
 import javax.lang.model.element.AnnotationMirror;
