+++ pom.xml
@@ -9,9 +9,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.annotationhelper</groupId>
-    <version>1.0.0-SNAPSHOT</version>
-    <artifactId>annotationhelper-aggregator</artifactId>
+    <groupId>org.mapstruct.annotations</groupId>
+    <version>0.5.0-SNAPSHOT</version>
+    <artifactId>annotations-aggregator</artifactId>
     <packaging>pom</packaging>
     <name>Annotation Helper Aggregator</name>
