+++ processor/pom.xml
@@ -9,13 +9,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.annotationhelper</groupId>
-        <version>1.0.0-SNAPSHOT</version>
-        <artifactId>annotationhelper-aggregator</artifactId>
+        <groupId>org.mapstruct.annotations</groupId>
+        <version>0.5.0-SNAPSHOT</version>
+        <artifactId>annotations-aggregator</artifactId>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>annotationhelper-processor</artifactId>
+    <artifactId>annotations-processor</artifactId>
     <packaging>jar</packaging>
     <name>Annotation Helper Processor</name>
 
@@ -23,7 +23,7 @@
 
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>annotationhelper-api</artifactId>
+            <artifactId>annotations-api</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
 
@@ -34,7 +34,7 @@
 
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>annotationhelper-test</artifactId>
+            <artifactId>annotations-test</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
         </dependency>
