+++ processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.annotationhelper.impl.GemsProcessor
+org.mapstruct.annotations.processor.GemProcessor
