+++ test/pom.xml
@@ -5,20 +5,20 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.annotationhelper</groupId>
-        <version>1.0.0-SNAPSHOT</version>
-        <artifactId>annotationhelper-aggregator</artifactId>
+        <groupId>org.mapstruct.annotations</groupId>
+        <version>0.5.0-SNAPSHOT</version>
+        <artifactId>annotations-aggregator</artifactId>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>annotationhelper-test</artifactId>
+    <artifactId>annotations-test</artifactId>
     <name>Annotation Helper Test</name>
 
     <dependencies>
 
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>annotationhelper-api</artifactId>
+            <artifactId>annotations-api</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
 
@@ -34,7 +34,7 @@
 <!--                    <annotationProcessorPaths>-->
 <!--                        <path>-->
 <!--                            <groupId>${project.parent.groupId}</groupId>-->
-<!--                            <artifactId>annotationhelper-processor</artifactId>-->
+<!--                            <artifactId>annotations-processor</artifactId>-->
 <!--                            <version>${project.parent.version}</version>-->
 <!--                        </path>-->
 <!--                        &lt;!&ndash; other annotation processors &ndash;&gt;-->
